requirements.txt
streamlit==1.20.0
plotly==5.13.1
pandas==1.4.3
numpy==1.23.1
folium==0.14.0
matplotlib==3.6.0
matplotlib-inline==0.1.6
streamlit-folium==0.11.1
Pillow==9.4.0
geopy==2.2.0